% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GlasshouseFileImport.R
\name{GlasshouseFileImport}
\alias{GlasshouseFileImport}
\title{Function to import DAT files from Creswick Glasshouses}
\usage{
GlasshouseFileImport(data)
}
\arguments{
\item{file}{Filename of the DAT file as character}
}
\value{
returns a data frame containing the data of the specified DAT file
}
\description{
Author: Markus Löw
Date: May 2015
Imports and processes a DAT file.
}
\examples{
GlasshouseFileImport("myfile.DAT")
}

