% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlasshouseFileImportPC2.R
\name{GlasshouseFileImportPC2}
\alias{GlasshouseFileImportPC2}
\title{Function to import DAT files from Creswick Glasshouses}
\usage{
GlasshouseFileImportPC2(data, glasshouse)
}
\arguments{
\item{data}{Filename of the DAT file as character}

\item{glasshouse}{Mandatory characterstring, either "PC2" or "teaching"}
}
\value{
returns a data frame containing the data of the specified DAT file
}
\description{
Author: Markus Löw
Date: May 2015
Imports and processes a DAT file.
}

