% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlasshouseFolderImport.R
\name{GlasshouseFolderImport}
\alias{GlasshouseFolderImport}
\title{An import function for all DAT glasshouse climate files from a folder}
\usage{
GlasshouseFolderImport(folder = ".", glasshouse)
}
\arguments{
\item{folder}{foldername of the DAT files. Defaults to current folder.}

\item{glasshouse}{Mandatory argument. Either "PC2" or "teaching" to specify which glasshouse the data come from.}
}
\value{
One data frame containing the data from all DAT files in the \code{foldername}
}
\description{
Author: Markus Löw
Date: May 2015
Imports and processes all Glasshouse DAT files from a folder. Uses \code{\link{GlasshouseFileImport}}.
}

