% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GlasshouseFolderImport.R
\name{GlasshouseFolderImport}
\alias{GlasshouseFolderImport}
\title{An import function for all DAT glasshouse climate files from a folder}
\usage{
GlasshouseFolderImport(folder = ".")
}
\arguments{
\item{folder}{Filename of the DAT file. Defaults to current folder
GlasshouseFolderImport(folder)}
}
\description{
Imports and processes all Glasshouse DAT files from a folder.
}

