% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GlasshouseFolderImport.R
\name{GlasshouseFolderImport}
\alias{GlasshouseFolderImport}
\title{An import function for all DAT glasshouse climate files from a folder}
\usage{
GlasshouseFolderImport(folder = ".")
}
\arguments{
\item{foldername}{of the DAT files. Defaults to current folder.}
}
\value{
One data frame containing the data from all DAT files in the \code{foldername}
}
\description{
Author: Markus Löw
Date: May 2015
Imports and processes all Glasshouse DAT files from a folder. Uses \code{\link{GlasshouseFileImport}}.
}
\examples{
GlasshouseFolderImport()
GlasshouseFolderImport("C:/Mydata/Climate")
}

